backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi
uvicorn
pydantic
pillow
python-multipart
python-dotenv
huggingface_hub
transformers
torch
torchvision
torchaudio
pdf2image
pytesseract
google-generativeai
